+++ packages/dataviews/CHANGELOG.md
@@ -12,6 +12,10 @@
 
 -   Remove some unused dependencies ([#62010](https://github.com/WordPress/gutenberg/pull/62010)).
 
+### Enhancement
+
+-   `label` prop in Actions API can be either a `string` value or a `function`, in case we want to use information from the selected items. ([#61942](https://github.com/WordPress/gutenberg/pull/61942)).
+
 ## 1.2.0 (2024-05-16)
 
 ### Internal